InvocationId is now a ushort. Addressed issue with InvocationIds wrapping around and re-using an active id.

+++ src/NexNet/Messages/InvocationMessage.cs
@@ -27,7 +27,7 @@ internal partial class InvocationMessage : IMessageBase, IInvocationMessage
     }
 
     [MemoryPackOrder(0)]
-    public int InvocationId { get; set; }
+    public ushort InvocationId { get; set; }
 
     [MemoryPackOrder(1)]
     public ushort MethodId { get; set; }
